renamed main page base files, added init db code
